+++ lib/china_regions.rb
@@ -1,6 +1,5 @@
 require "china_regions/version"
 require 'china_regions/engine' if defined? Rails
+require "china_regions/action_view/extension"
 
 module ChinaRegions;end
-
-require 'china_regions/helper'
