+++ lib/china_regions/engine.rb
@@ -1,5 +1,13 @@
 require "rails"
 
 module ChinaRegions
-  class Engine < Rails::Engine;end
+  if ::Rails.version > "3.1"
+    class Engine < ::Rails::Engine; end
+  else
+    class Railtie < ::Rails::Railtie
+      ActiveSupport.on_load(:action_view) do
+        ::ActionView::Base.send :include, Penmanship::ActionView::Extension
+      end  
+    end
+  end
 end
