+++ lib/china_regions/version.rb
@@ -1,3 +1,3 @@
 module ChinaRegions
-  VERSION = "0.0.1"
+  VERSION = "0.0.2"
 end
