+++ lib/tasks/china_regions.rake
